+++ changelog.md
@@ -27,6 +27,7 @@ Codion Change Log
 - FileInputPanel improved, related refactoring.
 - FilterTableCellRenderer.Builder.condition() renamed filter().
 - TableConditionPanel.conditionModel() renamed conditions().
+- FilterTable.Builder.resizeRowToFitEditor() added along with RESIZE_ROW_TO_FIT_EDITOR configuration value.
 ### is.codion.framework.model
 - DefaultEntityConditionModel, no longer wraps a single condition in a condition combination.
 - EntityConditionModel.AdditionalCondition added, related refactoring.